+++ frontend/appflowy_flutter/lib/plugins/ai_chat/presentation/chat_input/chat_input.dart
@@ -1,3 +1,4 @@
+import 'package:appflowy/plugins/ai_chat/application/chat_entity.dart';
 import 'package:appflowy/plugins/ai_chat/application/chat_file_bloc.dart';
 import 'package:appflowy/plugins/ai_chat/application/chat_input_action_bloc.dart';
 import 'package:appflowy/plugins/ai_chat/application/chat_input_action_control.dart';
