+++ frontend/appflowy_flutter/lib/plugins/ai_chat/presentation/chat_user_invalid_message.dart
@@ -1,4 +1,4 @@
-import 'package:appflowy/plugins/ai_chat/application/chat_bloc.dart';
+import 'package:appflowy/plugins/ai_chat/application/chat_entity.dart';
 import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart';
